feat(utils): add timezone metrics
