+++ packages/utils/src/metricsEncrypt.js
@@ -1,0 +1,11 @@
+// @ts-check
+
+/**
+ * Encodes given data with btoa
+ * @param {any} data
+ * @returns {string}
+ */
+const encryptMetric = data =>
+  btoa(unescape(encodeURIComponent(JSON.stringify(data))));
+
+export default encryptMetric;
