--- packages/utils/types/metrics.d.ts
@@ -1,9 +1,0 @@
-declare module '@endpass/utils/metrics' {
-  export function getExternalIP(): Promise<string | null>;
-
-  export function getFingerprintJsonData(): Promise<{
-    [key: string]: string | number;
-  }>;
-
-  export function encryptMetric(data: any): string;
-}
