+++ packages/utils/types/metricsEncrypt.d.ts
@@ -1,0 +1,3 @@
+declare module '@endpass/utils/metricsEncrypt' {
+  export default function encryptMetric(data: any): string;
+}
