+++ packages/utils/types/metricsFingerprint.d.ts
@@ -1,0 +1,5 @@
+declare module '@endpass/utils/metricsFingerprint' {
+  export default function getFingerprint(): Promise<{
+    [key: string]: string | number;
+  }>;
+}
