+++ packages/utils/types/metricsIP.d.ts
@@ -1,0 +1,3 @@
+declare module '@endpass/utils/metricsIP' {
+  export default function getIP(): Promise<string | null>;
+}
