+++ packages/utils/types/metricsTimezone.d.ts
@@ -1,0 +1,6 @@
+declare module '@endpass/utils/metricsTimezone' {
+  export default function getTimezone(): {
+    timezone: string;
+    gmt: number | string;
+  };
+}
